requirements.txt
langchain-core>=0.2.0
langchain-google-genai>=1.0.0
langgraph>=0.1.0
google-generativeai>=0.5.0
python-dotenv>=1.0.0
